feat: add password reset utils class to hold the utils method used in reset password workflow
